gemini_sorter/requirements.txt
python-dotenv
pathlib
requests
tqdm
argparse
